Confirmation when delete permission and fix table list SKP Tahunan

+++ resources/views/permissions/index.blade.php
@@ -32,7 +32,7 @@
                                         <td>
                                         <a href="{{ URL::to('permissions/'.$permission->id.'/edit') }}" class="btn btn-info pull-left" style="margin-right: 3px;">Edit</a>
 
-                                        {!! Form::open(['method' => 'DELETE', 'route' => ['permissions.destroy', $permission->id] ]) !!}
+                                        {!! Form::open(['method' => 'DELETE', 'route' => ['permissions.destroy', $permission->id], 'onsubmit' => 'return confirm("Yakin menghapus data ini ('. $permission->name . ')?")']) !!}
                                         {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                                         {!! Form::close() !!}
